backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy==2.0.48
python-dotenv==1.2.2
openai==2.30.0
httpx==0.28.1
pydantic==2.12.5
